+++ react/src/components/forms/MultiSelectDropDown/MultiSelectDropDown.stories.js
@@ -5,7 +5,7 @@ import { action } from '@storybook/addon-actions';
 
 import MultiSelectDropDown from './index';
 
-storiesOf('molecules', module)
+storiesOf('forms|molecules', module)
   .addDecorator(withKnobs({ escapeHTML: false }))
   .add('MultiSelectDropDown', (() => {
     const props = {
