+++ react/src/components/forms/MultiSelectDropDown/style.scss
