+++ react/src/components/forms/TypeAheadDropdown/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import ButtonWithIcon from '../../atoms/buttons/ButtonWithIcon';
-import InputTextTypeAhead from '../../forms/InputTextTypeAhead';
+import InputTextTypeAhead from '../InputTextTypeAhead';
 import Icon from '../../atoms/icons/Icon';
 import './style.css';
