+++ react/src/components/forms/TypeAheadDropdown/style.scss
